Create add task functionality
Program successfully adds tasks to the list and reorders the list. Moves the completed tasks to the bottom and the new task to the end of the current tasks. Need to fix bug when you close the task list, it does not add the buttons back to the screen even though the buttons were saved to a dictionary.